+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Support for all CSS color names
 
 ## [0.1.0] - 2025-07-21